pyproject.toml
[tool.poetry]
name = "informatics"
version = "0.1.0"
description = "teaching computer science in secondary schools"
authors = ["raqazhet <qazhetrazaq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
psycopg-pool = "^3.2.6"
psycopg = "^3.2.6"
django-model-utils = "^5.0.0"
django-autocomplete-light = "^3.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
